added tooltip and modal JS refacto
