+++ app/javascript/components/tooltip.js
@@ -1,0 +1,5 @@
+const demoTooltip = $(document).ready(function(){
+  $('[data-toggle="tooltip"]').tooltip();
+});
+
+export { demoTooltip };
