+++ app/views/demos/dashboard.html.erb
@@ -1,10 +1,11 @@
     <!-- Modal -->
     <%= render "modal-demo" %>
-      <div class="row mt-5 d-flex justify-content-center">
+      <div class="row mt-5 py-4 d-flex justify-content-center">
         <div class="col-7 col-lg-10">
           <div class="cards-demo">
-            <a type="button" id="buttonModal" class="btn btn-demo" data-toggle="modal" data-target="#demoModal">
-              <div class="circle">
+            <%# <a type="button" id="buttonModal" class="btn btn-demo" > %>
+             <a type="button" id="buttonModal" class="btn btn-demo" data-toggle="modal" data-target="#demoModal">
+              <div class="circle" data-toggle="tooltip" data-placement="bottom" title="Add a new demo">
                 <%= image_tag "icons/add.svg" %>
               </div>
             </a>
